refactor(login): Refactor login flow into single saga.
